RM 2356-natives - VTK API Updates

+++ src/edu/jhuapl/sbmt/tools/PerspectiveImagePreRenderer2.java
@@ -49,7 +49,7 @@ public class PerspectiveImagePreRenderer2
 
 		double[] boundingBox = footprints.get(0).GetBounds();
 		Just.of(renderableImage)
-			.operate(new OfflimbPlaneGeneratorOperators(offLimbFootprintDepth, smallBodyModels, boundingBox, footprints.get(0).GetNumberOfPoints()))
+			.operate(new OfflimbPlaneGeneratorOperators(offLimbFootprintDepth, smallBodyModels, boundingBox, (int)footprints.get(0).GetNumberOfPoints()))
 			.subscribe(PairSink.of(offLimbPolydata))
 			.run();
